fix float error

+++ src/estimator.py
@@ -23,8 +23,8 @@
 
 def estimator(data):
 
-    currentlyInfected = int(data["reportedCases"] * 10)
-    currentlyInfected_severe = int(data["reportedCases"] * 50)
+    currentlyInfected = float(int(data["reportedCases"] * 10))
+    currentlyInfected_severe = float(int(data["reportedCases"] * 50))
 
     infectionsByRequestedTime = currentlyInfected * 512
     infectionsByRequestedTime_severe = currentlyInfected_severe * 512
